#473 move artifact translation logic to dedicated artifacts module
